requirements.txt
Flask==3.1.2
python-gnupg==0.5.5
gunicorn==23.0.0